Make supported proj iteration a shared fixture
- Added units tests for coastlines
- Moved warning filter to pyproject
